split out test file
